+++ fe/fe-core/src/main/java/com/starrocks/analysis/FunctionCallExpr.java
@@ -964,56 +964,4 @@ public class FunctionCallExpr extends Expr {
         }
         return false;
     }
-
-    // geo function whose name starts_with 'st_' and is not 'st_astext' generates invalid utf8 strings. so it cannot
-    // be used as a argument to a string-typed parameter of a function.
-    // 1. select reverse(st_circle(...)) unacceptable
-    // 2. select reverse(st_astext(st_circle(...)) acceptable
-    // 3. select reverse(cast(st_circle(...) as varchar)) unacceptable
-    // 4. select reverse(cast(cast(st_circle(...) as varchar) as varchar))) unacceptable
-    public static void checkGeoFunctionGeneratedInvalidUtf8(FunctionCallExpr node) throws SemanticException {
-        Function fn = node.getFn();
-        String fnName = fn.functionName().toLowerCase();
-        if (fnName.equals(FunctionSet.ST_ASTEXT) || !(fn instanceof ScalarFunction)) {
-            return;
-        }
-        int numChildren = node.getChildren().size();
-        Type[] args = fn.getArgs();
-        for (int i = 0; i < numChildren; ++i) {
-            Type childType;
-            // variadic functions
-            if (i > args.length - 1) {
-                if (!fn.hasVarArgs()) {
-                    continue;
-                }
-                childType = fn.getVarArgsType();
-            } else {
-                childType = fn.getArgs()[i];
-            }
-
-            if (!childType.isStringType()) {
-                continue;
-            }
-
-            // strip cast(expr as varchar)
-            Expr child = node.getChild(i);
-            while (child instanceof CastExpr) {
-                Type grandsonType = child.getChild(0).getType();
-                if (grandsonType.isStringType()) {
-                    child = child.getChild(0);
-                } else {
-                    break;
-                }
-            }
-            if (child instanceof FunctionCallExpr) {
-                FunctionCallExpr fnCallChild = (FunctionCallExpr) child;
-                String name = fnCallChild.getFn().functionName().toLowerCase();
-                if (!name.equals(FunctionSet.ST_ASTEXT) && name.startsWith(FunctionSet.GEO_FUNCTION_PREFIX)
-                        && fnCallChild.getFn().getReturnType().isStringType()) {
-                    throw new SemanticException(String.format("Function '%s' cannot invoke '%s'", fn.functionName(), name));
-                }
-            }
-        }
-        return;
-    }
 }
